+++ packages/diagrams/frontend/sirius-components-diagrams/src/renderer/palette/DraggablePalette.types.ts
@@ -1,0 +1,29 @@
+/*******************************************************************************
+ * Copyright (c) 2026 Obeo.
+ * This program and the accompanying materials
+ * are made available under the terms of the Eclipse Public License v2.0
+ * which accompanies this distribution, and is available at
+ * https://www.eclipse.org/legal/epl-2.0/
+ *
+ * SPDX-License-Identifier: EPL-2.0
+ *
+ * Contributors:
+ *     Obeo - initial API and implementation
+ *******************************************************************************/
+import { PaletteExtensionSectionProps } from '@eclipse-sirius/sirius-components-palette';
+import { XYPosition } from '@xyflow/react';
+import { GQLPalette, GQLTool } from './Palette.types';
+
+export interface DraggablePaletteState {
+  controlledPosition: XYPosition;
+}
+
+export interface DraggablePaletteProps {
+  x: number;
+  y: number;
+  palette: GQLPalette;
+  diagramElementIds: string[];
+  onToolClick: (tool: GQLTool) => void;
+  onClose: () => void;
+  paletteToolListExtensions: React.ReactElement<PaletteExtensionSectionProps>[];
+}
